AAE-10187 review magic number list adding hexdecimal support
