+++ activiti-cloud-modeling-service/activiti-cloud-services-modeling/activiti-cloud-services-modeling-service/src/main/java/org/activiti/cloud/services/modeling/validation/magicnumber/FileContentValidatorConfiguration.java
@@ -1,6 +1,5 @@
 package org.activiti.cloud.services.modeling.validation.magicnumber;
 
-import org.activiti.cloud.services.modeling.validation.magicnumber.FileMagicNumber.FileMagicNumberList;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
