+++ activiti-cloud-modeling-service/activiti-cloud-services-modeling/activiti-cloud-services-modeling-service/src/main/resources/executable-magic-numbers.properties
@@ -1,22 +1,32 @@
-executable-filter.magic-number[0].name=Windows Executable
+executable-filter.magic-number[0].name=DOS MZ executable and its descendants
 executable-filter.magic-number[0].string=MZ
-executable-filter.magic-number[1].name=Windows Installer
-executable-filter.magic-number[1].bytes=-48, -49, 17, -32, -95, -79, 26, -31
-executable-filter.magic-number[2].name=Sh Type 1
-executable-filter.magic-number[2].string=#!/
-executable-filter.magic-number[3].name=Sh Type 2
-executable-filter.magic-number[3].string=#!\ /
-executable-filter.magic-number[4].name=Sh Type 3
-executable-filter.magic-number[4].string=#!\t/
-executable-filter.magic-number[5].name=Sh Type 4
-executable-filter.magic-number[5].string=eval &quot;exec
-executable-filter.magic-number[6].name=Linux (Elf)
-executable-filter.magic-number[6].bytes=0x7F, 0x45, 0x4C, 0x46
-executable-filter.magic-number[7].name=Mac OS Executable
-executable-filter.magic-number[7].bytes=-17, -65, -67, -17, -65, -67, -17, -65, -67, -17, -65, -67, 0, 0, 0, 2, 1, 0, 0
-executable-filter.magic-number[8].name=Mac OS DMG (zlib) Type 1
-executable-filter.magic-number[8].bytes=66, 90, 104, 57, 49, 65, 89, 38, 83, 89
-executable-filter.magic-number[9].name=Mac OS DMG (zlib) Type 2
-executable-filter.magic-number[9].bytes=120, -38, 99, 96, 24, 5, 67, 24, -4, -5, -1, -1, 29, 16, 51, 2, -103
-executable-filter.magic-number[10].name=Perl
-executable-filter.magic-number[10].bytes=35, 33, 47, 117, 115, 114, 47
+executable-filter.magic-number[1].name=DOS ZM executable and its descendants
+executable-filter.magic-number[1].string=ZM
+executable-filter.magic-number[2].name=Windows Installer
+executable-filter.magic-number[2].bytes=-48, -49, 17, -32, -95, -79, 26, -31
+executable-filter.magic-number[3].name=Bash script
+executable-filter.magic-number[3].string=#!/
+executable-filter.magic-number[4].name=Bash script
+executable-filter.magic-number[4].string=#!\ /
+executable-filter.magic-number[5].name=Bash script
+executable-filter.magic-number[5].string=#!\t/
+executable-filter.magic-number[6].name=Bash script
+executable-filter.magic-number[6].string=eval &quot;exec
+executable-filter.magic-number[7].name=Linux (Elf)
+executable-filter.magic-number[7].bytes=0x7F, 0x45, 0x4C, 0x46
+executable-filter.magic-number[8].name=Mach-O binary (32-bit)
+executable-filter.magic-number[8].string=0xfeedface
+executable-filter.magic-number[8].offset=4096
+executable-filter.magic-number[9].name=Mach-O binary (64-bit)
+executable-filter.magic-number[9].string=0xfeedfacf
+executable-filter.magic-number[9].offset=4096
+executable-filter.magic-number[10].name=Mach-O binary (reverse byte ordering scheme, 64-bit)
+executable-filter.magic-number[10].string=0xcffaedfe
+executable-filter.magic-number[10].offset=4096
+executable-filter.magic-number[11].name=Mach-O binary (reverse byte ordering scheme, 32-bit)
+executable-filter.magic-number[11].string=0xcefaedfe
+executable-filter.magic-number[11].offset=4096
+executable-filter.magic-number[12].name=Java class
+executable-filter.magic-number[12].string=0xcafebabe
+executable-filter.magic-number[13].name=Dalvik Executable
+executable-filter.magic-number[13].string=64, 65, 78, 0A, 30, 33, 35, 00
